+++ ModrinthAPIConnect/DELETE.py
@@ -1,0 +1,3 @@
+class DELETE:
+    def __init__(self):
+        pass
