+++ ModrinthAPIConnect/PATCH.py
@@ -1,0 +1,3 @@
+class Connect:
+    def __init__(self):
+        pass
