+++ ModrinthAPIConnect/__init__.py
@@ -1,0 +1,1 @@
+from ModrinthAPIConnect.GET import Project, Version, User
